Reorganize sections about messages and server/requestor behavior
Previously it was all mashed together, and during review was found to be fairly ambiguous. This change creates a new server behavior and requestor behavior section, and makes it clearer what the difference is between an initial update and a refresh. In particular, the fact that a server may not be able to distinguish between an initial update and a request is explictly acknowledged, and the implications discussed.